translate/requirements.txt
PySide6==6.7.1
sounddevice==0.4.7
numpy==1.26.4
soundfile==0.12.1
python-dotenv==1.0.1
faster-whisper==1.0.1
rich==13.7.1
protobuf==5.27.3
scipy==1.11.4
websocket-client==1.8.0
psutil==6.0.0
pycaw==20240210
comtypes==1.4.8
soundcard==0.4.3

